Move merge_settings into namespace detail
