+++ libvast/src/format/test.cpp
@@ -24,9 +24,10 @@
 #include "vast/error.hpp"
 #include "vast/factory.hpp"
 #include "vast/logger.hpp"
-#include "vast/settings.hpp"
 #include "vast/table_slice_builder.hpp"
 #include "vast/table_slice_builder_factory.hpp"
+
+#include <caf/settings.hpp>
 
 using caf::holds_alternative;
 using caf::visit;
